+++ package.json
@@ -20,6 +20,7 @@
     "@types/styled-components": "^5.1.34",
     "chart.js": "^4.4.3",
     "chartjs-plugin-datalabels": "^2.2.0",
+    "framer-motion": "^11.3.4",
     "html2canvas": "^1.4.1",
     "react": "^18.3.1",
     "react-chartjs-2": "^5.2.0",
