+++ src/main/mainstyles.tsx
@@ -148,88 +148,6 @@ export const OwnglyphFont = styled.h1`
   padding: 0 20px; // 양옆 패딩 추가
 `;
 
-// 닉네임 입력 관련 부분
-export const InputContainer = styled.div`
-  display: flex;
-  align-items: center;
-  background-color: #e0e0e0;
-  border-radius: 30px;
-  padding: 5px;
-  top: 80%; 
-  left: 50%;
-  transform: translate(-50%, -50%);
-  position: absolute;
-  width: 500px;
-  margin-top: 20px;
-  opacity: 1; 
-  z-index: 10; 
-
-  @media (max-width: 768px) {
-    width: 90%;
-  }
-
-  @media (max-width: 480px) {
-    width: 95%;
-  }
-`;
-
-export const StyledInput = styled.input`
-  border: none;
-  border-top-left-radius: 30px;
-  background-color: #e0e0e0; 
-  border-bottom-left-radius: 30px;
-  padding: 10px 20px;
-  font-size: clamp(14px, 3vw, 20px);
-  height: 25px; 
-  line-height: 40px; 
-  flex: 1;
-  outline: none;
-  font-family: 'GmarketSansMedium', sans-serif;
-
-  &::placeholder {
-    font-family: 'GmarketSansMedium', sans-serif;
-    font-size: clamp(14px, 3vw, 20px);
-  }
-`;
-
-export const StyledButton = styled.button`
-  background-color: #555;
-  color: white;
-  border: none;
-  border-radius: 25px;
-  width: 80px;
-  height: 40px;
-  cursor: pointer;
-  font-size: clamp(12px, 2.5vw, 16px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  opacity: 1; 
-  z-index: 10;
-  transform: translateX(-10px); 
-  line-height: 40px; 
-  
-  &:hover {
-    background-color: #777;
-  } 
-  
-  span {
-    display: inline-block;
-    text-align: center;
-    width: 100%; 
-    margin-top: 3px;
-  }
-
-  @media (max-width: 768px) {
-    width: 60px;
-    height: 35px;
-  }
-
-  @media (max-width: 480px) {
-    width: 50px;
-    height: 30px;
-  }
-`;
 
 // 온보딩 3P 사진 돌아가는 css
 // 캐릭터 회전 애니메이션을 위한 키프레임 추가
